requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlmodel==0.0.24
psycopg2-binary==2.9.6
dotenv==0.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.10.1